[Silver I] Title: 기타 레슨, Time: 8 ms, Memory: 2412 KB -BaekjoonHub

+++ 백준/Silver/2343. 기타 레슨/기타 레슨.cc
@@ -5,21 +5,21 @@ using namespace std;
 
 int main() {
     ios::sync_with_stdio(0), cin.tie(0), cout.tie(0);
-    long long n, m;
+    int n, m;
     cin >> n >> m;
 
-    long long left = 1;
-    long long right = 1000000000;
+    int left = 1;
+    int right = 1000000000;
 
-    vector<long long> classes(n, 0);
-    for (long long i = 0; i < n; i++) {
+    vector<int> classes(n, 0);
+    for (int i = 0; i < n; i++) {
         cin >> classes[i];
         left = max(left, classes[i]);
     }
     while (left <= right) {
-        long long mid = (left + right) / 2;
-        long long sumLength = 0;
-        long long count = 0;
+        int mid = (left + right) / 2;
+        int sumLength = 0;
+        int count = 0;
     
         for (const auto length : classes) {
             if (sumLength + length <= mid)
